requirements.txt
streamlit
tweepy
openai
python-dotenv
pandas